+++ packages/node/src/node/ServerNode.ts
@@ -4,6 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+import type { ClusterBehavior } from "#behavior/cluster/ClusterBehavior.js";
 import { ActionContext } from "#behavior/context/ActionContext.js";
 import { CommissioningServer } from "#behavior/system/commissioning/CommissioningServer.js";
 import { ControllerBehavior } from "#behavior/system/controller/ControllerBehavior.js";
@@ -15,13 +16,36 @@ import { SessionsBehavior } from "#behavior/system/sessions/SessionsBehavior.js"
 import { SubscriptionBehavior } from "#behavior/system/subscription/SubscriptionBehavior.js";
 import { Endpoint } from "#endpoint/Endpoint.js";
 import type { Environment } from "#general";
-import { Construction, DiagnosticSource, Identity, MatterError, asyncNew, errorOf } from "#general";
-import { FabricManager, Interactable, OccurrenceManager, ServerInteraction, SessionManager } from "#protocol";
+import {
+    asyncNew,
+    Construction,
+    createPromise,
+    deepCopy,
+    DiagnosticSource,
+    errorOf,
+    Identity,
+    isObject,
+    Logger,
+    MatterError,
+    MaybePromise,
+} from "#general";
+import {
+    Fabric,
+    FabricManager,
+    Interactable,
+    OccurrenceManager,
+    ServerInteraction,
+    SessionManager,
+    Val,
+} from "#protocol";
+import { ObjectSchema } from "#types";
 import { RootEndpoint as BaseRootEndpoint } from "../endpoints/root.js";
 import { Node } from "./Node.js";
 import { ClientNodes } from "./client/ClientNodes.js";
 import { ServerEnvironment } from "./server/ServerEnvironment.js";
 import { ServerNodeStore } from "./storage/ServerNodeStore.js";
+
+const logger = Logger.get("ServerNode");
 
 /**
  * Thrown when there is an error during factory reset.
@@ -180,13 +204,104 @@ export class ServerNode<T extends ServerNode.RootEndpoint = ServerNode.RootEndpo
     protected override async initialize() {
         await ServerEnvironment.initialize(this);
 
+        this.env.get(FabricManager).events.deleted.on(fabric => this.#sanitizeFabricData(fabric));
+
         await super.initialize();
     }
 
     /**
-     * By default on factory reset we erase all stored data.
+     * Sanitize Fabric scoped data and events when a fabric is removed.
+     * The logic walks through all endpoints and removes relevant fabric-scoped attribute values for the relevant fabric
+     * from the state. After all state changes are processed, it removes all occurrences of fabric-scoped events.
+     */
+    async #sanitizeFabricData(fabric: Fabric) {
+        const fabricIndex = fabric.fabricIndex;
+        const fabricRelevantEvents = new Set<string>();
+        const stateUpdatePromises = new Array<Promise<void>>();
+        let sanitizedAttributes = 0;
+        await this.visit(async endpoint => {
+            for (const type of Object.values(endpoint.behaviors.supported)) {
+                const cluster = (type as ClusterBehavior.Type)?.cluster;
+                if (!cluster) {
+                    continue;
+                }
+                const stateUpdate = {} as Val.Struct;
+                const elements = endpoint.behaviors.elementsOf(type);
+                if (elements.attributes.size) {
+                    // Iterate over all attributes and check if they are fabric-scoped
+                    for (const attributeName of elements.attributes) {
+                        const attr = cluster.attributes[attributeName];
+                        if (attr.fabricScoped) {
+                            const value = (endpoint.stateOf(type) as Val.Struct)[attributeName];
+                            // If the value contains data for the fabric being removed, remove the data
+                            if (Array.isArray(value) && value.length > 0) {
+                                const filtered = deepCopy(value).filter(entry => entry.fabricIndex !== fabricIndex);
+                                if (filtered.length !== value.length) {
+                                    sanitizedAttributes++;
+                                    stateUpdate[attributeName] = filtered;
+                                }
+                            }
+                        }
+                    }
+                    // If we have any state updates for this behavior, we need to set the state.
+                    // Errors are being logged and ignored
+                    if (Object.keys(stateUpdate).length > 0) {
+                        const { resolver, promise } = createPromise<void>();
+                        (endpoint.eventsOf(type) as any).stateChanged?.on(resolver);
+                        try {
+                            await endpoint.setStateOf(type, stateUpdate);
+                            stateUpdatePromises.push(promise);
+                        } catch (error) {
+                            logger.warn(
+                                `Could not sanitize fabric-scoped attributes for cluster ${cluster.name} on endpoint ${endpoint.id}`,
+                                error,
+                            );
+                        }
+                    }
+                }
+                if (elements.events.size) {
+                    // If we have events also check if they are fabric scoped and collect them
+                    for (const eventName of elements.events) {
+                        const event = cluster.events[eventName];
+                        if ((event.schema as ObjectSchema<any>).isFabricScoped) {
+                            fabricRelevantEvents.add(`${cluster.id}-${event.id}`);
+                        }
+                    }
+                }
+            }
+        });
+
+        // Wait for all state changed to be executed before processing events
+        await Promise.allSettled(stateUpdatePromises);
+
+        let sanitizedEvents = 0;
+        // Now we can remove all occurrences of fabric-scoped events when payload is bound to this fabric index
+        if (fabricRelevantEvents.size > 0) {
+            const occurrences = this.env.get(OccurrenceManager);
+            for await (const event of occurrences.get()) {
+                if (
+                    fabricRelevantEvents.has(`${event.clusterId}-${event.eventId}`) &&
+                    isObject(event.payload) &&
+                    event.payload.fabricIndex === fabricIndex
+                ) {
+                    // Remove occurrences of fabric-scoped events
+                    const result = occurrences.remove(event.number);
+                    if (MaybePromise.is(result)) {
+                        await result;
+                    }
+                    sanitizedEvents++;
+                }
+            }
+        }
+        logger.info(
+            `Sanitized ${sanitizedAttributes} attributes and ${sanitizedEvents} events because of fabric removal (index ${fabricIndex})`,
+        );
+    }
+
+    /**
+     * By default, on factory reset we erase all stored data.
      *
-     * If this is inappropriate for your application you may override to alter the behavior.   Matter requires that all
+     * If this is inappropriate for your application, you may override to alter the behavior.  Matter requires that all
      * "security- and privacy-related data and key material" is removed on factory reset.
      *
      * @see {@link MatterSpecification.v12.Core} § 13.4